+++ server/src/modules/uploadMulter/uploadMulter.ts
@@ -1,0 +1,18 @@
+import path from "node:path";
+import multer from "multer";
+
+const storage = multer.diskStorage({
+  destination: (req, file, cb) => {
+    cb(null, "public/uploads/products");
+  },
+  filename: (req, file, cb) => {
+    const timestamp = Date.now();
+    const ext = path.extname(file.originalname);
+    const basename = path.basename(file.originalname, ext);
+    cb(null, `produit-${timestamp}-${basename}${ext}`);
+  },
+});
+
+const upload = multer({ storage });
+
+export const productImagesUpload = upload.any();
